# docs/swagger.yml
openapi: 3.0.0
info:
  title: "Resume API"
  version: "1.0.0"
  description: "Serverless Role Based Auth API powered by JWT"
  contact:
    email: "sai@pathuri.net"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "https://resume.saipathuri.me"
paths:
  /about:
    get:
      summary: "Retrieve the about section of the resume."
      responses:
        "200":
          description: "About section retrieved."
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/About"
        "401":
          description: "Authentication failed."
    put:
      summary: "Update the about section of the resume."
      requestBody:
        required: true
        description: "About information to update."
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/About"
      responses:
        "200":
          description: "Successfully updated."
        "401":
          description: "Authentication failed."
...
components:
  schemas:
    About:
      type: object
      properties:
        fullName:
          type: string
        address:
          type: object
          properties:
            lineOne:
              type: string
            lineTwo:
              type: string
        phoneNumber:
          type: string
        email:
          type: string
        tagLine:
          type: string
...